portal/requirements.txt
asgiref==3.2.4
Django==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
